func of modal
